pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flask-apcore"
version = "0.3.0"
description = "Flask Extension for apcore AI-Perceivable Core integration, exposes Flask routes as MCP tools with auto-discovery, schema inference, YAML binding, and built-in observability. Supports MCP server (stdio/HTTP) for LLM tool integration."
requires-python = ">=3.11"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "aipartnerup", email = "tercel.yi@gmail.com" },
]
keywords = [
    "flask",
    "apcore",
    "mcp",
    "llm",
    "ai",
    "tools",
    "pydantic",
    "yaml",
    "observability",
    "schema",
    "module",
    "extension",
    "cli",
    "auto-discovery",
    "binding",
    "integration"
]
dependencies = [
    "flask>=3.0",
    "apcore>=0.13.0",
    "apcore-toolkit>=0.2.0",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
smorest = ["flask-smorest>=0.42"]
restx = ["flask-restx>=1.0"]
mcp = ["apcore-mcp>=0.10.0"]
dev = [
    "pytest>=7.0",
    "pytest-flask>=1.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "coverage>=7.0",
    "apdev[dev]>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/flask_apcore"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py311"
line-length = 120

[tool.apdev]
base_package = "flask_apcore"

[tool.mypy]
python_version = "3.11"
strict = true

[project.urls]
Homepage = "https://github.com/aipartnerup/"
Repository = "https://github.com/aipartnerup/flask-apcore"
Documentation = "https://github.com/aipartnerup/flask-apcore#readme"
Issues = "https://github.com/aipartnerup/flash-apcore/issues"